Bug Fix and Added Features
So previously there was a bug that when we ran the code, sometimes the replay would run multiple times, or sometimes when we try to replay our recorded code twice it wouldn't replay the second time. We fixed that bug. Also added vectors for recording lift motors. Still need to add comments to this code.