Update main.cpp
Changed the flick motor function. It was overheating because the motor wasn't properly installed, but also because we were making the motor go forward and reverse repeatedly a lot of times in an instant, forcing it to resist itself, so I put a stop function in-between every movement to solve the problem.